Added another enemy and a death state. Try doing a run on the stage
BaseEnemy.cs - Fixed a possible null reference

Checkpoint.cs - renamed the animator variable to match the style we used with player and enemy.

Damager.cs - Used on things such as spikes or blades to damage the player when they touch it.

FlyMan.cs - New enemy that hovers between jumps. (I would like to rename him).

GameManager.cs - Broke up the functionality for falling and reviving.

Player.cs - Added a dead state. Renamed the bools to be a statement.

SpikeMan.cs - removed the Start override since we weren't using it.